ai_service/requirements.txt
flask
flask-cors
openai
python-dotenv
requests
numpy
pandas
scikit-learn
textblob
nltk
spacy
transformers
torch
gunicorn
python-dateutil
geopy
beautifulsoup4
lxml